refactor: move libstack to root
