+++ composebinary/utils_test.go
