+++ composebinary/wrapper_test.go
