+++ libstack.go
